Rename reducers and change imports in root reducer
