+++ src/reducers/errorReducer.js
@@ -1,4 +1,4 @@
-export const invalidUserReducer = (state = null, action) => {
+export const errorReducer = (state = null, action) => {
     switch (action.type) {
       case 'INVALID_USER':
         return "Please login to save an album.";
