+++ src/reducers/favoritesReducer.js
